+++ packages/connectkit/src/components/Pages/Solana/PayWithSolanaToken/index.tsx
@@ -13,7 +13,6 @@ import {
   buildCheckoutPayload,
   checkoutPayment,
   FeeResponseData,
-  FeeType,
   formatPaymentResponseToHydratedOrder,
   getCanonicalDestination,
   getChainExplorerTxUrl,
@@ -23,11 +22,12 @@ import {
   solana,
   WalletPaymentOption,
 } from "@rozoai/intent-common";
+import { formatUnits } from "viem";
 import { useContactSupport } from "../../../../hooks/useContactSupport";
 import { useRozoPay } from "../../../../hooks/useRozoPay";
 import { ROZO_EVENTS } from "../../../../lib/analytics/events";
 import { useAnalytics } from "../../../../provider/AnalyticsProvider";
-import { getCachedFee, resolveOrderAppId } from "../../../../utils/feeCache";
+import { buildFeeQuoteParams, getCachedFee } from "../../../../utils/feeCache";
 import Button from "../../../Common/Button";
 import PaymentBreakdown from "../../../Common/PaymentBreakdown";
 import TokenLogoSpinner from "../../../Spinners/TokenLogoSpinner";
@@ -178,20 +178,26 @@ const PayWithSolanaToken: React.FC = () => {
 
         // @NOTE: Fee calculation
         const destToken = currentOrder.destFinalCallTokenAmount?.token;
+        const destAmountAtomic = currentOrder.destFinalCallTokenAmount?.amount;
+        const toUnits = destAmountAtomic && destToken
+          ? formatUnits(BigInt(destAmountAtomic), destToken.decimals)
+          : option.required.usd.toString();
         setFeeLoading(true);
-        const feeData = await getCachedFee({
-          appId: resolveOrderAppId(currentOrder, paymentState.payParams?.appId),
-          type: paymentState.payParams?.feeType ?? FeeType.ExactIn,
-          sourceChainId: option.required.token.chainId.toString(),
-          sourceTokenSymbol: option.required.token.symbol,
-          amount: option.required.usd.toString(),
-          destChainId: destToken.chainId.toString(),
-          destReceiverAddress:
-            getCanonicalDestination(currentOrder).finalDestinationAddress ??
-            paymentState.payParams?.toAddress ??
-            "",
-          destTokenSymbol: destToken.symbol,
-        });
+        const feeData = await getCachedFee(
+          buildFeeQuoteParams({
+            order: currentOrder,
+            payParams: paymentState.payParams,
+            destChainId: destToken.chainId,
+            destTokenAddress: destToken.token,
+            destAddress:
+              getCanonicalDestination(currentOrder).finalDestinationAddress ??
+              "",
+            sourceChainId: option.required.token.chainId,
+            sourceTokenAddress: option.required.token.token,
+            toUnits,
+            feeUsd: option.fees.usd,
+          }),
+        );
         setFeeLoading(false);
 
         if (feeData.error) {